evaluation/requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==4.5
accelerate==1.7.0
adapter-transformers==4.0.0
adapters==1.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bzip2==1.0.8
ca-certificates==2025.4.26
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.3.0
executing==2.2.0
faiss-gpu==1.7.2
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.2
fsspec==2025.3.0
gdown==5.2.0
google-auth==2.40.3
google-genai==1.19.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.1
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.0
keyutils==1.6.1
kiwisolver==1.4.8
krb5==1.21.3
ld_impl_linux-64==2.43
libedit==3.1.20250104
libexpat==2.7.0
libffi==3.4.6
libgcc==15.1.0
libgcc-ng==15.1.0
libgomp==15.1.0
liblzma==5.8.1
libnsl==2.0.1
libsodium==1.0.20
libsqlite==3.49.2
libstdcxx==15.1.0
libstdcxx-ng==15.1.0
libuuid==2.38.1
libxcrypt==4.4.36
libzlib==1.3.1
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
ncurses==6.5
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openssl==3.5.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.1
prv-accountant==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
pysocks==1.7.1
python==3.10.17
python-dateutil==2.9.0.post0
python_abi==3.10
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
readline==8.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack_data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tk==8.6.13
tokenizers==0.21.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
trl==0.18.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
xgboost==3.0.2
xxhash==3.5.0
yarl==1.20.0
zeromq==4.3.5
zipp==3.22.0
